+++ doc/changelog.rst
@@ -2,7 +2,9 @@ Changelog
 =========
 * :support:`4` Add encryption and decryption for caesar cipher.
 * :support:`8` Add contributing guidelines.
+* :support:`6` Add transposition cipher decryption.
 * :support:`3` Add Travis build status, code coverage, and Read The
   Docs badges.
 * :support:`2` Add Travis CI build script.
 * :feature:`1` Created initial project folder structure.
+